requirements.txt
absl-py==0.9.0
asgiref==3.2.10
asn1crypto==1.3.0
astor==0.8.0
astunparse==1.6.3
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
cycler==0.10.0
Django==3.0.7
django-cors-middleware==1.5.0
django-oauth-toolkit==1.3.2
django-oauth2-provider==0.2.6.1
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-oauth==1.1.0
enum-compat==0.0.3
future==0.18.2
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.14.1
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
joblib==0.15.1
Keras==2.1.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.16.0
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
opt-einsum==3.1.0
pandas==1.0.5
Pillow @ file:///tmp/build/80754af9/pillow_1594307298074/work
protobuf==3.12.3
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy @ file:///tmp/build/80754af9/scipy_1592930540905/work
seaborn==0.10.1
shortuuid==1.0.1
simplejson==3.17.0
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.3.1
torchfm==0.7.0
torchtext==0.6.0
torchvision==0.2.0
tqdm==4.46.1
uritemplate==3.0.1
urllib3==1.25.9
utils==1.0.1
Werkzeug==0.16.1
wrapt==1.12.1
